update swg controller to handle entitlements prompt on article use case (#69)
# Feature Description
Google have asked that we add entitlements checking to Article pages. 
Currently the `defaultOnwardsEntitlementsJourney` function in `n-swg` defines the properties for the 'you already have an account via google' overlay and then shows the overlay. 
We want to show this message on articles and will use `n-messaging-client` to do that, however to implement this properly in `n-messaging-client` we need to use `o-banner` rather than the overlay created by `n-swg`.
This PR splits out the `defaultOnwardsEntitlementsJourney` function into two:
 - `getEntitledOnwardJourneyProps` which defines and returns the properties for the button on the entitlements prompt. This can be used agnostic of whether the `n-swg` overlay or an `n-messaging-client` banner is used as the format for the prompt 
 - `defaultOnwardsEntitlementsJourney` - uses the properties returned by `getEntitledOnwardJourneyProps` to render the `n-swg` overlay

# Link to Ticket / Card:
https://trello.com/c/WgGJ6ueb

## Image
![image](https://user-images.githubusercontent.com/17846996/44850603-a8439880-ac55-11e8-923d-7b22f8f7d811.png)

## To do

- [x] button styling per SIWG - being covered in different card - https://trello.com/c/WCQm5hVx
- [x] write test for new function

## Related PRs

- https://github.com/Financial-Times/n-messaging-client/pull/111